requirements.txt
tensorflow-rocm
onnx
tf2onnx
numpy<2.0.0
